attempt day 7

+++ day_07/src/utils/parser.rs
@@ -8,7 +8,7 @@ static REGEX: LazyLock<Regex> = LazyLock::new(|| {
     Regex::new(r"(\d+): (.+)").expect("Unable to create regex")
 });
 
-pub fn parse_input(file_path: &str) -> Result<HashMap<isize, Vec<isize>>, BridgeRepairError> {
+pub fn parse_input(file_path: &str) -> Result<HashMap<i64, Vec<i64>>, BridgeRepairError> {
     // define input and regex
     let input = fs::read_to_string(file_path)?;
     let re = &REGEX;
@@ -19,10 +19,10 @@ pub fn parse_input(file_path: &str) -> Result<HashMap<isize, Vec<isize>>, Bridge
     // iterate over lines in input
     for line in input.lines() {
         for cap in re.captures_iter(line) {
-            let key: isize = cap[1].parse()?;
-            let value: Vec<isize> = cap[2]
+            let key: i64 = cap[1].parse()?;
+            let value: Vec<i64> = cap[2]
             .split_whitespace()
-            .map(|x| x.parse::<isize>())
+            .map(|x| x.parse::<i64>())
             .collect::<Result<Vec<_>, _>>()?;
 
             result.insert(key, value);
